requirements.txt
python-dotenv
openai
pydantic
anthropic
google-generativeai